trading_system/requirements.txt
# Professional Trading System v2.0 - Dependencies

# Core dependencies
httpx>=0.24.0
asyncio-mqtt>=0.11.0
pyyaml>=6.0
numpy>=1.24.0

# API and web server
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Data processing
pandas>=2.0.0
aiofiles>=23.0.0

# Web dashboard dependencies
jinja2>=3.1.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0